Exclude platforms with no native package support when testing native package support

+++ client/test/test_dependency.rb
@@ -633,20 +633,24 @@ class TpkgDependencyTests < Test::Unit::TestCase
     # a tpkg with the same name. For this we need a native package that is
     # generally available on systems that developers are likely to use, I'm
     # going to use wget for now.
-    nativedep = make_package(:output_directory => @tempoutdir, :change => { 'name' => 'nativedep' }, :dependencies => {'wget' => {'native' => true}}, :remove => ['operatingsystem', 'architecture'])
-    tpkgdep = make_package(:output_directory => @tempoutdir, :change => { 'name' => 'tpkgdep' }, :dependencies => {'wget' => {}}, :remove => ['operatingsystem', 'architecture'])
-    wget = make_package(:output_directory => @tempoutdir, :change => { 'name' => 'wget' }, :remove => ['operatingsystem', 'architecture'])
-    parent = make_package(:output_directory => @tempoutdir, :change => { 'name' => 'parent' }, :dependencies => {'nativedep' => {}, 'tpkgdep' => {}}, :remove => ['operatingsystem', 'architecture'])
-    Dir.mktmpdir('testbase') do |testbase|
-      mixeddeppkgs = [nativedep, tpkgdep, wget, parent]
-      tpkg = Tpkg.new(:base => testbase, :sources => mixeddeppkgs)
-      solution_packages = tpkg.best_solution([{:name => 'parent', :type => :tpkg}], {}, ['parent'])
-      # The solution should include the four tpkgs plus a native wget
-      assert_equal(5, solution_packages.length)
-      assert(solution_packages.any? {|sp| sp[:metadata][:name] == 'wget' && (sp[:source] == :native_available || sp[:source] == :native_installed)})
-      assert(solution_packages.any? {|sp| sp[:metadata][:name] == 'wget' && sp[:source].include?('wget-1.0-1.tpkg')})
-      mixeddeppkgs.each do |mdp|
-        FileUtils.rm_f(mdp)
+    # FIXME: should have a better way of excluding platforms with no native
+    # package support
+    if Tpkg.get_os !~ /windows/
+      nativedep = make_package(:output_directory => @tempoutdir, :change => { 'name' => 'nativedep' }, :dependencies => {'wget' => {'native' => true}}, :remove => ['operatingsystem', 'architecture'])
+      tpkgdep = make_package(:output_directory => @tempoutdir, :change => { 'name' => 'tpkgdep' }, :dependencies => {'wget' => {}}, :remove => ['operatingsystem', 'architecture'])
+      wget = make_package(:output_directory => @tempoutdir, :change => { 'name' => 'wget' }, :remove => ['operatingsystem', 'architecture'])
+      parent = make_package(:output_directory => @tempoutdir, :change => { 'name' => 'parent' }, :dependencies => {'nativedep' => {}, 'tpkgdep' => {}}, :remove => ['operatingsystem', 'architecture'])
+      Dir.mktmpdir('testbase') do |testbase|
+        mixeddeppkgs = [nativedep, tpkgdep, wget, parent]
+        tpkg = Tpkg.new(:base => testbase, :sources => mixeddeppkgs)
+        solution_packages = tpkg.best_solution([{:name => 'parent', :type => :tpkg}], {}, ['parent'])
+        # The solution should include the four tpkgs plus a native wget
+        assert_equal(5, solution_packages.length)
+        assert(solution_packages.any? {|sp| sp[:metadata][:name] == 'wget' && (sp[:source] == :native_available || sp[:source] == :native_installed)})
+        assert(solution_packages.any? {|sp| sp[:metadata][:name] == 'wget' && sp[:source].include?('wget-1.0-1.tpkg')})
+        mixeddeppkgs.each do |mdp|
+          FileUtils.rm_f(mdp)
+        end
       end
     end
